Move OsuCursor resize logic to OsuCursorContainer
